move OutputRecorder
