+++ protostar/commands/test/test_case_runners/test_case_runner.py
@@ -9,8 +9,8 @@ from protostar.commands.test.test_environment_exceptions import (
     BreakingReportedException,
     ReportedException,
 )
-from protostar.commands.test.test_output_recorder import OutputRecorder
 from protostar.test_runner.stopwatch import Stopwatch
+from protostar.test_runner.test_output_recorder import OutputRecorder
 from protostar.test_runner.test_results import (
     BrokenTestCaseResult,
     FailedTestCaseResult,
