+++ protostar/commands/test/test_result_formatter.py
@@ -2,7 +2,6 @@ from pathlib import Path
 from typing import Callable, Dict, List
 
 from protostar.commands.test.test_environment_exceptions import ExceptionMetadata
-from protostar.commands.test.test_output_recorder import OutputName, format_output_name
 from protostar.protostar_exception import UNEXPECTED_PROTOSTAR_ERROR_MSG
 from protostar.test_runner import (
     BrokenFuzzTestCaseResult,
@@ -10,11 +9,13 @@ from protostar.test_runner import (
     BrokenTestSuiteResult,
     FailedFuzzTestCaseResult,
     FailedTestCaseResult,
+    OutputName,
     PassedFuzzTestCaseResult,
     PassedTestCaseResult,
     SkippedTestCaseResult,
     TestResult,
     UnexpectedBrokenTestSuiteResult,
+    format_output_name,
 )
 from protostar.utils.log_color_provider import log_color_provider
